# feat: 35% faster decompression with less boundary check