@a10y 2024-10-03
Thanks @XiangpengHao ! Just two small comments (perhaps I'm being overly paranoid about `add` vs `byte_add` but just prefer to be explicit).