@XiangpengHao 2024-10-03
Thanks for the review! I've updated the code and rerun the benchmark and see no perf difference (just to be sure).